apps/backend/pyproject.toml
[project]
name = "jewelryai-backend"
version = "0.1.0"
description = "FastAPI gateway for JewelryAI MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "trimesh>=4.7.0",
  "python-multipart>=0.0.20",
  "pillow>=11.2.0"
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "httpx>=0.28.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
